[AOT Refactor] input/output analysis
[ghstack-poisoned]
