+++ torch/_functorch/_aot_autograd/schemas.py
@@ -11,6 +11,10 @@ from torch._subclasses.fake_tensor import is_fake
 
 from .. import config
 
+from .utils import strict_zip
+
+zip = strict_zip
+
 OutputType = Enum(
     "OutputType",
     (
